move `createStatementData` to a new file
